requirements.txt
aiokafka==0.8.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.103.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kafka-python==2.0.2
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
python-decouple==3.8
python-telegram-bot==20.5
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
